requirements.txt
# Core dependencies
pyqlib>=0.9.0
pandas>=1.5.0
numpy>=1.23.0,<2.0.0
scipy>=1.9.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Technical Analysis
TA-Lib>=0.4.24
pandas-ta>=0.3.14b

# Data Sources
yfinance>=0.2.0
newsapi-python>=0.2.7
sec-edgar-downloader>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Scheduling & Background Tasks
schedule>=1.2.0
apscheduler>=3.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
redis>=4.5.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# MLOps
mlflow>=2.5.0
optuna>=3.2.0

# Monitoring
prometheus-client>=0.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.65.0
click>=8.1.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
asyncio>=3.4.3

# Data processing
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0  # For testing FastAPI

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
notebook>=6.5.0

# Deployment
gunicorn>=21.0.0
python-json-logger>=2.0.7

# Social Media (optional)
tweepy>=4.14.0  # Twitter
praw>=7.7.0     # Reddit
